+++ NEWS.md
@@ -11,6 +11,7 @@ New features
 Bug fixes and minor improvements
 --------------------------------
 
+* ``cat_piece()`` now has very basic support for fudge/fate dice (#139)
 * Fixes bug when launching command-line version of `view_game()`
 
 ppgames 0.7.0